Feat: create Form component and refactor Login and Signup to use it
