+++ src/components/Home.js
@@ -2,12 +2,12 @@ import React from 'react';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
 
-const Home = ({ loggedInStatus, handleLogout, history }) => {
+const Home = (props) => {
   const handleClick = () => {
     axios.delete('http://localhost:3001/logout', { withCredentials: true })
       .then(response => {
-        handleLogout();
-        history.push('/');
+        props.handleLogout();
+        props.history.push('/');
       })
       .catch(error => console.log(error));
   };
@@ -19,7 +19,7 @@ const Home = ({ loggedInStatus, handleLogout, history }) => {
       <Link to="/signup">Sign Up</Link>
       <br />
       {
-          { loggedInStatus }
+          props.loggedInStatus
             ? <Link to="/logout" onClick={handleClick}>Log Out</Link>
             : null
         }
