+++ src/components/registrations/Login.js
@@ -1,115 +1,13 @@
-import React, { Component } from 'react';
-import axios from 'axios';
-import { Link } from 'react-router-dom';
+import React from 'react';
+import Form from '../Form';
 
-class Login extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      username: '',
-      email: '',
-      password: '',
-      errors: '',
-    };
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.redirect = this.redirect.bind(this);
-    this.handleErrors = this.handleErrors.bind(this);
-  }
+const Login = props => {
+  const { history, handleLogin } = props;
+  const type = 'login';
 
-  componentWillMount() {
-    return this.props.loggedInStatus ? this.redirect() : null
-  }
-
-  handleChange(event) {
-    const { name, value } = event.target;
-    this.setState({
-      [name]: value,
-    });
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    const { username, email, password } = this.state;
-    const user = {
-      username,
-      email,
-      password,
-    };
-
-    axios.post('http://localhost:3001/login', { user }, { withCredentials: true })
-      .then(response => {
-        if (response.data.logged_in) {
-          this.props.handleLogin(response.data);
-          this.redirect();
-        } else {
-          this.setState({
-            errors: response.data.errors,
-          });
-        }
-      })
-      .catch(error => console.log('api errors:', error));
-  }
-
-  redirect() {
-    this.props.history.push('/');
-  }
-
-  handleErrors() {
-    return (
-      <div>
-        <ul>
-          {this.state.errors.map(error => <li key={error}>{error}</li>)}
-        </ul>
-      </div>
-    );
-  }
-
-  render() {
-    const { username, email, password } = this.state;
-    return (
-      <div>
-        <h1>Log In</h1>
-        {' '}
-        <form onSubmit={this.handleSubmit}>
-          <input
-            placeholder="username"
-            type="text"
-            name="username"
-            value={username}
-            onChange={this.handleChange}
-          />
-          <input
-            placeholder="email"
-            type="text"
-            name="email"
-            value={email}
-            onChange={this.handleChange}
-          />
-          <input
-            placeholder="password"
-            type="password"
-            name="password"
-            value={password}
-            onChange={this.handleChange}
-          />
-          <button placeholder="submit" type="submit">
-            Log In
-          </button>
-          <div>
-            or
-            {' '}
-            <Link to="/signup">sign up</Link>
-          </div>
-        </form>
-        <div>
-          {
-            this.state.errors ? this.handleErrors() : null
-          }
-        </div>
-      </div>
-    );
-  }
-}
+  return (
+    <Form type={type} history={history} handleLogin={handleLogin} />
+  );
+};
 
 export default Login;
