+++ src/constants.js
@@ -1,0 +1,7 @@
+const SIGNUP_URL = 'http://localhost:3001/users';
+const LOGIN_URL = 'http://localhost:3001/login';
+
+export {
+  SIGNUP_URL,
+  LOGIN_URL,
+};
